Added solution to StudentRepository exc in lab 6`
