+++ lab6/academia/CMakeLists.txt
@@ -1,8 +1,8 @@
 
 project(studentrepository)
 
-set(HEDEAR_FILES StudentRepository.h StudyYear.h)
-set(SOURCE_FILES StudentRepository.cpp StudyYear.cpp)
+set(HEDEAR_FILES StudentRepository.h StudyYear.h Student.h)
+set(SOURCE_FILES StudentRepository.cpp StudyYear.cpp Student.cpp)
 
 add_library(libstudentrepository STATIC ${SOURCE_FILES})
 target_include_directories(libstudentrepository PUBLIC .)
@@ -10,4 +10,4 @@ target_include_directories(libstudentrepository PUBLIC .)
 add_executable(studentrepository ${SOURCE_FILES} main.cpp)
 
 target_link_libraries(studentrepository libstudentrepository)
-
+target_link_libraries(libstudentrepository libstudentrepositoryqueries)
