+++ lab6/academia/StudentRepository.h
@@ -7,29 +7,13 @@
 #include <string>
 #include <iostream>
 #include <vector>
+#include <StudentRepositoryQueries.h>
 #include "StudyYear.h"
+#include "Student.h"
 
 using ::std::string;
 
 namespace academia{
-    class Student{
-    public:
-        Student(string id = "", string first_name = "", string last_name = "", string program = "", StudyYear year = 3);
-        const bool operator==(const Student &other)const;
-        void ChangeFirstName(const string new_name);
-        void ChangeLastName(const string new_name);
-        const string Id(void)const;
-        const string FirstName(void)const;
-        const string LastName(void)const;
-        const string Program(void)const;
-        const StudyYear Year(void)const;
-    private:
-        string id_;
-        string first_name_;
-        string last_name_;
-        string program_;
-        StudyYear year_;
-    };
 
     class StudentRepository{
     public:
@@ -37,6 +21,7 @@ namespace academia{
         Student& operator[](const string cmp);
         unsigned long StudentCount(void)const;
         const bool operator==(const StudentRepository &other)const;
+        std::vector<Student> FindByQuery(const Query &query) const;
     private:
         std::vector<Student> repo_;
     };
